requirements.txt
Django==5.0.2
django-cors-headers==4.3.1
requests==2.31.0
openai==1.12.0
python-dotenv==1.0.1
